Empirical prisoner's dilemma games from Nay and Vorobeychik

A dataset containing 168,386 total rounds of play in 30 different variations on the iterated prisoner's dilemma games. The data comes from J.J. Nay and Y. Vorobeychik, "Predicting Human Cooperation," PLOS ONE 11(5), e0155656 (2016).

NV_games

Format

A data frame with 168,386 rows and 51 variables:

period

Which turn of the given game

my.decision

The player's move in this turn

risk

Boolean variable: 1 indicates stochastic payoffs, 0 deterministic payoffs

delta

Probability the game ends after each round

r1

Normalized difference in payoff between both players cooperating and both defecting

r2

Normalized difference in payoff between both players cooperating and the payoff for being a sucker (cooperating when the opponent defects)

error

Probability that the player's intended move is switched to the opposite move

data

Which dataset did this game come from: AM = Andreoni & Miller; BR = Bereby-Meyer & Roth; DB = Dal Bo; DF = Dal Bo & Frechette; DO = Duffy & Ochs; FO = Friedman & Oprea; FR = Fudenberg, Rand, & Dreber; and KS = Kunreuther, Silvasi, Bradlow & Small

my.decision1

The player's move in the previous turn

my.decision2

The player's move two turns ago

my.decision3

The player's move three turns ago

my.decision4

The player's move four turns ago

my.decision5

The player's move five turns ago

my.decision6

The player's move six turns ago

my.decision7

The player's move seven turns ago

my.decision8

The player's move eight turns ago

my.decision9

The player's move nine turns ago

other.decision1

The opponent's move in the previous turn

other.decision2

The opponent's move two turns ago

other.decision3

The opponent's move three turns ago

other.decision4

The opponent's move four turns ago

other.decision5

The opponent's move five turns ago

other.decision6

The opponent's move six turns ago

other.decision7

The opponent's move seven turns ago

other.decision8

The opponent's move eight turns ago

other.decision9

The opponent's move nine turns ago

my.payoff1

The player's payoff in the previous turn

my.payoff2

The player's payoff two turns ago

my.payoff3

The player's payoff three turns ago

my.payoff4

The player's payoff four turns ago

my.payoff5

The player's payoff five turns ago

my.payoff6

The player's payoff six turns ago

my.payoff7

The player's payoff seven turns ago

my.payoff8

The player's payoff eight turns ago

my.payoff9

The player's payoff nine turns ago

other.payoff1

The opponent's payoff in the previous turn

other.payoff2

The opponent's payoff two turns ago

other.payoff3

The opponent's payoff three turns ago

other.payoff4

The opponent's payoff four turns ago

other.payoff5

The opponent's payoff five turns ago

other.payoff6

The opponent's payoff six turns ago

other.payoff7

The opponent's payoff seven turns ago

other.payoff8

The opponent's payoff eight turns ago

other.payoff9

The opponent's payoff nine turns ago

r

Reward: payoff when both players cooperate

t

Temptation: payoff when player defects and opponent cooperates

s

Sucker: Payoff when player cooperates and opponent defects

p

Punishment: payoff when both players defect

infin

Boolean: 1 indicates infinite game with probability delta of ending at each round; 0 indicates pre-determined number of rounds

contin

Boolean: 1 indicates the game is played in continuous time; 0 indicates discrete rounds

group

Which group (version of the game) is being played?

Source

doi: 10.1371/journal.pone.0155656